+++ src/components/AuthoritiesLookup/AuthoritiesLookup.js
@@ -20,12 +20,12 @@ import {
   AuthoritiesSearchContext,
   AuthoritiesSearchPane,
   AuthorityShape,
-  FILTERS,
   searchResultListColumns,
   SearchResultsList,
   SelectedAuthorityRecordContext,
   navigationSegments,
   useAutoOpenDetailView,
+  AUTH_REF_TYPES,
 } from '@folio/stripes-authority-components';
 
 import MarcAuthorityView from '../MarcAuthorityView';
@@ -52,9 +52,7 @@ const propTypes = {
   totalRecords: PropTypes.number.isRequired,
 };
 
-const excludedFilters = new Set([
-  FILTERS.REFERENCES,
-]);
+const excludedFilters = new Set([]);
 
 const AuthoritiesLookup = ({
   authorities,
@@ -98,9 +96,11 @@ const AuthoritiesLookup = ({
 
   const formatter = {
     [searchResultListColumns.LINK]: authority => {
-      const { id } = authority;
+      const { id, authRefType } = authority;
 
-      if (!id) return null;
+      if (!id || authRefType !== AUTH_REF_TYPES.AUTHORIZED) {
+        return null;
+      }
 
       if (isLinkingLoading && authorityIdToLink === id) {
         return (