+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Change history for ui-plugin-find-authority
 
-## [1.1.0] (IN PROGRESS)
+## [2.1.0] IN PROGRESS
+
+* [UIPFAUTH-44](https://issues.folio.org/browse/UIPFAUTH-44) Show an error message in modal after linking.
+
+## [2.0.0] (https://github.com/folio-org/ui-plugin-find-authority/tree/v2.0.0) (2023-02-24)
 
 * [UIPFAUTH-34](https://issues.folio.org/browse/UIPFAUTH-34) The pagination buttons are missing when the user changes the scale of the screen
 * [UIPFAUTH-36](https://issues.folio.org/browse/UIPFAUTH-36) Default search/browse option and Authority source file selections based on MARC bib field to be linked
 * [UIPFAUTH-37](https://issues.folio.org/browse/UIPFAUTH-37) bump stripes to 8.0.0 for Orchid/2023-R1
-* [UIPFAUTH-44](https://issues.folio.org/browse/UIPFAUTH-44) Show an error message in modal after linking.
+* [UIPFAUTH-47](https://issues.folio.org/browse/UIPFAUTH-47) Show References filter.
 
 ## [1.0.1] (https://github.com/folio-org/ui-plugin-find-authority/tree/v1.0.1) (2022-11-28)